Workspace Launcher Linux version 1 Ready
This commit completes an initial Linux version
of WorkSpace Launcher. It implements the logic
for user interfacing in Launcher.sh, and opens
saved workspace files, launches programs, sets
location and size of windows. Functionality is
broken up, and there's even some documentation

Among the remaining work is refactoring how it
obtains a newly launched program's PID. At the
moment, it returns the PID of the command that
launches the program, which does not match the
PID of the program itself. Once launched, that
command finishes and the program window cannot
be resized/relocated

Additionally, the fetched window's coordinates
appear to be a relative/inaccurate location as
they are <2x,2y>. Halving them corrects my use
case, but is unlikely to work consistently